+++ src/test/java/site/nomoreparties/stellarburgers/UserLoginTest.java
@@ -2,7 +2,12 @@ package site.nomoreparties.stellarburgers;
 
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
+import site.nomoreparties.stellarburgers.api.client.UserClient;
+import site.nomoreparties.stellarburgers.api.model.User;
+import site.nomoreparties.stellarburgers.api.util.UserGenerator;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
@@ -10,13 +15,21 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.equalToIgnoringCase;
 
 public class UserLoginTest {
+    private User user;
+    private UserClient userClient;
+
+    @Before
+    public void setUp() {
+        userClient = new UserClient();
+        user = UserGenerator.getRandom();
+    }
+
     @Test
     @DisplayName("Успешная авторизация пользователя")
     public void userCanBeLogin() {
-        UserClient userClient = new UserClient();
-        User user = UserGenerator.getRandom();
-
         Response createResponse = userClient.create(user);
+        String userToken = createResponse.path("accessToken");
+        user.setAccessToken(userToken);
 
         User userLogin = new User(user.getEmail(), user.getPassword());
 
@@ -32,19 +45,14 @@ public class UserLoginTest {
                 .body("refreshToken", is(notNullValue()))
                 .body("user.email", equalToIgnoringCase(user.getEmail()))
                 .body("user.name", equalToIgnoringCase(user.getName()));
-
-
-        String userToken = createResponse.path("accessToken");
-        userClient.delete(userToken);
     }
 
     @Test
     @DisplayName("Авторизация пользователя с неверным email")
     public void userLoginWithWrongEmail() {
-        UserClient userClient = new UserClient();
-        User user = UserGenerator.getRandom();
-
         Response createResponse = userClient.create(user);
+        String userToken = createResponse.path("accessToken");
+        user.setAccessToken(userToken);
 
         User userLogin = new User(user.getEmail() + "ru", user.getPassword());
 
@@ -57,19 +65,14 @@ public class UserLoginTest {
                 .statusCode(401)
                 .body("success", equalTo(false))
                 .body("message", equalTo("email or password are incorrect"));
-
-
-        String userToken = createResponse.path("accessToken");
-        userClient.delete(userToken);
     }
 
     @Test
     @DisplayName("Авторизация пользователя с неверным паролем")
     public void userLoginWithWrongPassword() {
-        UserClient userClient = new UserClient();
-        User user = UserGenerator.getRandom();
-
         Response createResponse = userClient.create(user);
+        String userToken = createResponse.path("accessToken");
+        user.setAccessToken(userToken);
 
         User userLogin = new User(user.getEmail(), user.getPassword() + "jfnD");
 
@@ -82,9 +85,20 @@ public class UserLoginTest {
                 .statusCode(401)
                 .body("success", equalTo(false))
                 .body("message", equalTo("email or password are incorrect"));
+    }
 
+    @After
+    @DisplayName("Удаление пользователя")
+    public void deleteUser() {
+        userClient = new UserClient();
+        String userToken = user.getAccessToken();
 
-        String userToken = createResponse.path("accessToken");
-        userClient.delete(userToken);
+        if (userToken != null) {
+            Response deleteResponse = userClient.delete(userToken);
+            deleteResponse
+                    .then()
+                    .assertThat()
+                    .statusCode(202);
+        }
     }
 }
